+++ packages/admin-sdk/src/ui/modal/index.ts
@@ -2,14 +2,21 @@ import { createSender } from '../../channel';
 
 export const open = createSender('uiModalOpen');
 export const close = createSender('uiModalClose');
+export const update = createSender('uiModalUpdate');
 
 export type uiModalOpen =
  {
    responseType: void,
-   title: string,
-   // Use the locationId to render the modal in the iframe
+   title?: string,
+   /**
+    * If a locationId is provided, the modal will
+    * be rendered and the content renders the location.
+    */
    locationId?: string,
-   // Use the textContent when no locationId is provided
+   /**
+    * If no locationId is provided, the textContent will be rendered in the modal.
+    * Only simple text is supported.
+    */
    textContent?: string,
    variant?: 'default'|'small'|'large'|'full',
    showHeader?: boolean,
@@ -27,6 +34,21 @@ export type uiModalClose =
     */
    locationId?: string,
  }
+
+/**
+ * Update the modal with the given locationId.
+ * Can be used to update the modal from every location
+ * after it has been opened.
+ */
+export type uiModalUpdate = {
+  responseType: void,
+  locationId: string,
+  title?: string,
+  showHeader?: boolean,
+  showFooter?: boolean,
+  closable?: boolean,
+  buttons?: buttonProps[],
+}
 
 export type buttonProps = {
   method: () => void,