+++ packages/component-library/src/components/entity/mt-entity-data-table/mt-entity-data-table.interfaces.ts
@@ -1,0 +1,5 @@
+export interface AvailableFilter {
+  id: string;
+  label: string;
+  filterType: "multi-select" | "boolean";
+}