+++ packages/component-library/src/components/form/_internal/mt-select-base/_internal/mt-select-selection-list.vue
@@ -266,10 +266,6 @@ export default defineComponent({
 
     clearSearchTerm() {
       this.inputInFocus = false;
-
-      if (this.searchTerm.length > 0) {
-        this.$emit("search-term-change", "");
-      }
     },
 
     onKeyDownDelete() {